Move abstract checks and result to core directory
