+++ colin/core/checks/abstract_check.py
