+++ colin/core/checks/images.py
